# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: EZScreen REST & Inter-Service API Specification
  description: Complete OpenAPI specification for EZScreen Core API Gateway and Internal Microservices
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/josh-ac8/EZScreen_API_specifications/1.0.0
  - url: https://api.ezscreen.io/api/v1
    description: Production Core API Gateway
  - url: http://localhost:8000/api/v1
    description: Local Development Core API
  - url: http://ai-core-services:8002/internal/v1
    description: Internal AI Core Microservice (Pure Read-Only Stateless Engine)
...
paths:
...
  /jobs:
    post:
      tags: [Job Descriptions]
      summary: Save & publish job description record
      operationId: createJobDescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, parsed_jd]
              properties:
                title:
                  type: string
                  example: Senior Java Developer
                description:
                  type: string
                  example: Looking for a Senior Java Developer...
                job_type:
                  type: string
                  enum: [part_time, full_time, contract]
                  example: full_time
                work_type:
                  type: string
                  enum: [onsite, hybrid, remote]
                  example: hybrid
                location:
                  type: string
                  example: Bangalore
                experience_min:
                  type: integer
                  example: 3
                experience_max:
                  type: integer
                  example: 5
                skills:
                  type: string
                  example: Java, Spring Boot, PostgreSQL, Docker
                status:
                  type: string
                  enum: [draft, published, closed]
                  example: published
                parsed_jd:
                  type: object
      responses:
        '201':
          description: Job description created and saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionResponse'
    get:
      tags: [Job Descriptions]
      summary: List job descriptions for organization
      operationId: listJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, closed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of job descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDescriptionResponse'
  /jobs/{id}:
    get:
      tags: [Job Descriptions]
      summary: View job description details
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job description detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionResponse'
    put:
      tags: [Job Descriptions]
      summary: Unified endpoint to update job description fields, parsed_jd requirements, and/or status
      operationId: updateJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Lead Java Backend Engineer
                description:
                  type: string
                  example: Updated job description text...
                job_type:
                  type: string
                  enum: [part_time, full_time, contract]
                work_type:
                  type: string
                  enum: [onsite, hybrid, remote]
                location:
                  type: string
                experience_min:
                  type: integer
                experience_max:
                  type: integer
                skills:
                  type: string
                status:
                  type: string
                  enum: [draft, published, closed]
                  example: closed
                parsed_jd:
                  type: object
      responses:
        '200':
          description: Job description updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionResponse'

  # ---------------------------------------------------------------------------
  # 4. PUBLIC CANDIDATE PORTAL
  # ---------------------------------------------------------------------------
  /public/jobs:
    get:
      tags: [Public Candidate Portal]
      summary: Candidate browse published jobs for organization subdomain
      operationId: listPublicJobs
      parameters:
        - name: org_subdomain
          in: query
          schema:
            type: string
            example: acme
      responses:
        '200':
          description: Published jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDescriptionResponse'
  /public/jobs/{id}:
    get:
      tags: [Public Candidate Portal]
      summary: Candidate view published job details
      operationId: getPublicJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Public job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDescriptionResponse'
...
components:
  schemas:
...
    JobDescriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 444e4567-e89b-12d3-a456-426614174000
        organization_id:
          type: string
          format: uuid
          example: 987e6543-e89b-12d3-a456-426614174000
        created_by:
          type: string
          format: uuid
          example: 333e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          example: Senior Java Developer
        description:
          type: string
          example: Looking for a Senior Java Developer...
        job_type:
          type: string
          enum: [part_time, full_time, contract]
          example: full_time
        work_type:
          type: string
          enum: [onsite, hybrid, remote]
          example: hybrid
        location:
          type: string
          example: Bangalore
        experience_min:
          type: integer
          example: 3
        experience_max:
          type: integer
          example: 5
        skills:
          type: string
          example: Java, Spring Boot, PostgreSQL, Docker
        status:
          type: string
          enum: [draft, published, closed]
          example: published
        parsed_jd:
          type: object
...